UMCS-479 (change) Move properties of charge class regarding account information to a trait and make it available for authorize transactions as well.
