+++ src/Resources/TransactionTypes/Authorization.php
@@ -27,6 +27,7 @@ namespace UnzerSDK\Resources\TransactionTypes;
 use UnzerSDK\Adapter\HttpAdapterInterface;
 use UnzerSDK\Exceptions\UnzerApiException;
 use UnzerSDK\Resources\Payment;
+use UnzerSDK\Traits\HasAccountInformation;
 use UnzerSDK\Traits\HasCancellations;
 use UnzerSDK\Traits\HasInvoiceId;
 use UnzerSDK\Traits\HasRecurrenceType;
@@ -37,6 +38,7 @@ class Authorization extends AbstractTransactionType
     use HasCancellations;
     use HasInvoiceId;
     use HasRecurrenceType;
+    use HasAccountInformation;
 
     /** @var float $amount */
     protected $amount = 0.0;
